Add cors configuration and change purchase office logic
Add cors configuration into /config/SecurityConfig.java
Return correct game state
Add exception handling for ml service
Change mofifier type column name

+++ backend/app/src/main/java/com/startupgame/dto/game/DeveloperCounts.java
@@ -1,4 +1,4 @@
 package com.startupgame.dto.game;
 
 
-public record DeveloperCounts(Integer juniors, Integer middles, Integer seniors){}
+public record DeveloperCounts(Long juniors, Long middles, Long seniors) { }
